hard coded refresh
Hard coded a 2 second refresh rate for the database statistics
displayed.  Hopefully handled onResume(), onPause(), onStop(), and
onDestroy() properly.  Used 3 values for the enum on handling what the
state of execution is, but because of the limited nature of the Thread
class implemented in Dalvik virtual machine and if I understand things
correctly, this probably only needs to be a boolean value.