fix error handling from thread
When errors were thrown from the background thread, a Toast was
attempting to show the error message.  But since this was happening from
the background thread, the application crashed trying to show the Toast.
Created some more variables so that errors can be safely shown to the
user, then pause the background thread otherwise the Toast will display
over and over again.