support 8.2
Well, support 8.2 where it can, and don't crash where it can't.
Background writer stats don't exist in 8.2 so that activity will display
a message saying so.  Some database stats exist in 8.2 but not as many
as in later versions.  So display what we can and 0 out the others.  It
may be better to not display stats that don't exist in 8.2 but that is
saved for later.